+++ core/src/test/java/com/bumble/appyx/routingsource/spotlight/operation/Routing.kt
@@ -1,0 +1,7 @@
+package com.bumble.appyx.routingsource.spotlight.operation
+
+internal sealed class Routing {
+    object Routing1 : Routing()
+    object Routing2 : Routing()
+    object Routing3 : Routing()
+}